requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-tinymce==3.1.0
gunicorn==20.0.4
importlib-metadata==2.0.0
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.4.1
whitenoise==5.2.0
zipp==3.4.0
